Finished Improving Jump and Movement - Also Begun Adding Interactibles
Jumping now has a functional input buffer, coyote time states and anti-gravity apexing. It also has a speed limiter so the player can't reach terminal velocity and can be further expanded with more features on the player falling and potentially even a double jump. Jump also now has tap and hold capability so you can hold space for longer or shorter to get a more accurate jump. The hold is also a massive boost. The colliders have also been improved, along with the camera no longer being directly linked with the player and the input only being checked when necessary.